Prioritize communication from tanks
This helps mitigate the problem where a tanks threat updates might be delayed because of other addons communicating. There are more improvements that can be done to the addon communication like discarding old updates that are still enqueued in ChatThrottleLib, but that would require doing changes in that library. Another thing to consider that could be done right now is monitor the communication delay using callbacks and notify the user if the delay is significant so that they can call out their threat manually in comms or turn off other addons.